Chapter 11

+++ Chapter11/src/q12/Main.java
@@ -7,15 +7,15 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.print("Enter 5 numbers: ");
-        ArrayList<Integer> arr = new ArrayList<>();
+        ArrayList<Double> arr = new ArrayList<>();
         for (int i = 0; i < 5; i++)
-            arr.add(scanner.nextInt());
+            arr.add(scanner.nextDouble());
         System.out.println(sum(arr));
     }
 
-    public static double sum(ArrayList<Integer> arr) {
+    public static double sum(ArrayList<Double> arr) {
         double total = 0;
-        for (Integer a : arr) {
+        for (Double a : arr) {
             total += a;
         }
         return total;
